refactor: 将 gateway aggregator 拆分为子模块并添加全局 deferred 上限和独立 retry
